correção do ID

+++ models.py
@@ -1,11 +1,12 @@
-from sqlalchemy import Column, Integer, String, Text
+from sqlalchemy import Column, String, Text, BigInteger
 from database import Base
+
 
 
 class User(Base):  # Renomeado para manter consistência com o users.py
     __tablename__ = "usuarios"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True) 
     name = Column(String)
     email = Column(String, unique=True, index=True)
     role = Column(String)
